refactor: Modify or_else_throw to throw custom error
BREAKING CHANGE: Optional.or_else_throw() now throws a custom error
instead of OptionalError. The new version of the method takes a
SupplyFunc and throws an error produced by the supply function if no
value is present. If the supply function is not specified, the method
throws an OptionalError.NOT_PRESENT error if no value is present.